apagar comunidades funcionando de novo

+++ webapp/src/main/java/com/brasfi/webapp/service/ComunidadeService.java
@@ -4,7 +4,9 @@ import com.brasfi.webapp.entities.Comunidade;
 import com.brasfi.webapp.repositories.ComunidadeRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.List; 
+import java.util.List;
+import java.util.Optional; // Importe Optional
+
 @Service
 public class ComunidadeService {
     private final ComunidadeRepository comunidadeRepository;
@@ -20,10 +22,14 @@ public class ComunidadeService {
     public List<Comunidade> listarTodasComunidades() {
         return comunidadeRepository.findAll();
     }
+
     @Transactional
-    public void excluirComunidade(Long id) {
-
-
-        comunidadeRepository.deleteById(id); 
+    public boolean excluirComunidade(Long id) {
+        // Verifica se a comunidade existe antes de tentar apagar
+        if (comunidadeRepository.existsById(id)) {
+            comunidadeRepository.deleteById(id);
+            return true; // Comunidade encontrada e excluída com sucesso
+        }
+        return false; // Comunidade não encontrada
     }
 }
